Crate drav

drav is a package of dimensionality reduction algorithms, with an emphasis on visualization.

Modules

util

Structs

BHTSNE

PCA

TSNE

TSNEOptions

Enums

PCATarget

Traits

NonParametricModel

Functions

load_mnist_images
Return a matrix representation of the set of images in the dataset file. Each row is an image, whose elements are the pixel-wise intensities laid out in row-major order.

load_mnist_labels

Type Definitions

Matrix

Vector