Struct drav::PCA

pub struct PCA {
    // some fields omitted
}

Methods

impl PCA

fn new(data: &Matrix<f32>, target: PCATarget) -> Result<PCA, PCAError>

fn transform_data<D>(&self, data: &ArrayBase<D, Ix2>) -> Matrix<f32> where D: Data<Elem=f32>

Transform a set of data into its reduced dimensionas

fn transform_datum<D>(&self, datum: &ArrayBase<D, Ix>) -> Vector<f32> where D: Data<Elem=f32>

Transform an individual vector.

fn reconstruct_datum(&self, datum: &Vector<f32>) -> Vector<f32>

Reconstruct the best approximation of the original vector using the transformed vector.

fn reconstruct_datum_partial(&self, datum: &Vector<f32>, dim: usize) -> Vector<f32>

Reconstruct the best approximatio of the original vector using a subset of the elements.

fn eigenvectors(&self) -> &Matrix<f32>

Reconstruct the best approximation for a data set using the transformed data set. Return the matrix of eigenvectors, with each column representing an eigenvector.

fn dim(&self) -> usize